Issue 3161: Display more information about in-house medications
