+++ apps/ehr/src/features/css-module/components/medication-administration/statuses/MedicationStatusChip.tsx
@@ -2,10 +2,9 @@ import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
 import { Box, Chip, MenuItem, Popover, Stack, Typography } from '@mui/material';
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { MedicationOrderStatusesType } from 'utils';
+import { MedicationOrderStatusesType, medicationStatusDisplayLabelMap } from 'utils';
 import { ExtendedMedicationDataForResponse } from 'utils';
 import { useMedicationManagement } from '../../../hooks/useMedicationManagement';
-import { medicationStatusDisplayLabelMap } from '../medication-editable-card/utils';
 
 interface MedicationStatusChipProps {
   medication?: ExtendedMedicationDataForResponse;
